airspec: Add text diff
